516 query update (#520)
* #516, materialize to the "additional_inputs" directory, using the current UserWorkspace

* timestamp to show err msg has changed

* da_college_debt files

* django version update

* updated webpack
