+++ tworaven_common_apps/datamart_endpoints/datamart_util_isi.py
@@ -1,18 +1,20 @@
 import json
-import zipfile
-from io import BytesIO
+from tworaven_apps.utils.json_helper import json_loads, json_dumps
 
 from django.conf import settings
-import pandas as pd
+from tworaven_apps.user_workspaces.utils import get_latest_user_workspace
+
 from tworaven_apps.data_prep_utils.new_dataset_util import NewDatasetUtil
 from tworaven_apps.user_workspaces.models import UserWorkspace
 from tworaven_apps.configurations.utils import get_latest_d3m_config
+from tworaven_apps.utils.random_info import get_timestamp_string_readable
 from tworaven_apps.utils.basic_response import (ok_resp,
                                                 err_resp)
+from tworaven_apps.utils.dict_helper import (clear_dict,)
 from tworaven_common_apps.datamart_endpoints.static_vals import \
     (cached_response,
      cached_response_baseball)
-from tworaven_common_apps.datamart_endpoints.info_util import \
+from tworaven_common_apps.datamart_endpoints.datamart_info_util import \
     (get_isi_url,
      get_nyu_url)
 
@@ -53,7 +55,7 @@ class DatamartJobUtilISI(object):
 
     @staticmethod
     def datamart_get_metadata(data):
-
+        """Use endpoint: /new/get_metadata_single_file"""
         try:
             response = requests.post(
                 get_isi_url() + '/new/get_metadata_single_file',
@@ -98,13 +100,35 @@ class DatamartJobUtilISI(object):
         return ok_resp(response['data'])
 
     @staticmethod
-    def datamart_search(query, data_path=None, limit=None):
+    def datamart_search(query_str, data_path=None, limit=None):
         # TODO disable #debug
         # return ok_resp(json.loads(cached_response_baseball))
         # TODO: respect limit
         limit = 100
 
-        files = {'query': ('query.json', query)}
+        query_info_json = json_loads(query_str)
+        if not query_info_json.success:
+            user_msg = ('There is something wrong with the search parameters.'
+                        ' (expected a JSON string)')
+            return err_resp(user_msg)
+
+        query_dict = query_info_json.result_obj
+
+        clear_dict(query_dict)
+        if not query_dict:
+            no_params_msg = ('There are no search parameters.'
+                             ' You must have at least 1.')
+            return err_resp(no_params_msg)
+
+        formatted_json_info = json_dumps(query_dict)
+        if not formatted_json_info.success:
+            return err_resp('Failed to convert query to JSON. %s' % \
+                            formatted_json_info.err_msg)
+
+        print(f'formatted query: {formatted_json_info.result_obj}')
+        files = {'query': ('query.json', formatted_json_info.result_obj)}
+
+        #files = {'query': ('query.json', query)}
         if data_path and os.path.exists(data_path):
             files['file'] = open(data_path, 'r')
 
@@ -156,7 +180,8 @@ class DatamartJobUtilISI(object):
         #print('processed_datasets', processed_datasets)
 
         if not processed_datasets:
-            return err_resp('No datasets found.')
+            return err_resp('No datasets found. (%s)' % \
+                            (get_timestamp_string_readable(),))
 
         # Normally, the data is sorted by score in descending order,
         # but just in case...
@@ -171,22 +196,21 @@ class DatamartJobUtilISI(object):
 
 
     @staticmethod
-    def datamart_materialize(search_result):
+    def datamart_materialize(user_workspace, search_result):
+        """Materialize the dataset"""
+        if not isinstance(user_workspace, UserWorkspace):
+            return err_resp('user_workspace must be a UserWorkspace')
 
-        d3m_config = get_latest_d3m_config()
-        if not d3m_config:
-            user_msg = 'datamart_materialize failed. no d3m config'
-            LOGGER.error(user_msg)
-            return err_resp(user_msg)
         datamart_id = search_result['datamart_id']
-        materialize_folderpath = os.path.join(d3m_config.temp_storage_root,
-                                              'materialize',
-                                              str(datamart_id))
+        materialize_folderpath = os.path.join(\
+                                user_workspace.d3m_config.additional_inputs,
+                                'materialize',
+                                str(datamart_id))
 
+        print('materialize_folderpath', materialize_folderpath)
         if os.path.exists(materialize_folderpath):
             response = None
         else:
-
             try:
                 response = requests.get(\
                                 get_isi_url() + '/new/materialize_data',
@@ -199,7 +223,11 @@ class DatamartJobUtilISI(object):
             if response['code'] != "0000":
                 return err_resp(response['message'])
 
-        return ok_resp(DatamartJobUtilISI.save(materialize_folderpath, response))
+        save_info = DatamartJobUtilISI.save_datamart_file(materialize_folderpath, response)
+        if not save_info.success:
+            return err_resp(save_info.err_msg)
+
+        return ok_resp(save_info.result_obj)
 
     @staticmethod
     def datamart_augment(user_workspace, data_path, search_result, left_columns, right_columns, exact_match=False, **kwargs):
@@ -257,7 +285,7 @@ class DatamartJobUtilISI(object):
                             'augment',
                             str(datamart_id))
 
-        save_info = DatamartJobUtilISI.save(augment_folderpath, response)
+        save_info = DatamartJobUtilISI.save_datamart_file(augment_folderpath, response)
         if not save_info.success:
             return err_resp(save_info.err_msg)
 
@@ -277,7 +305,8 @@ class DatamartJobUtilISI(object):
 
 
     @staticmethod
-    def save(folderpath, response):
+    def save_datamart_file(folderpath, response):
+        """Save materialize response as a file"""
         data_filepath = os.path.join(folderpath, 'tables', 'learningData.csv')
 
         if os.path.exists(folderpath):
